refactor: change styled-component to sass
