+++ frontend/declarations.d.ts
@@ -2,3 +2,7 @@ declare module "*.svg" {
   const content: string;
   export default content;
 }
+declare module "*.module.scss" {
+  const classes: { [key: string]: string };
+  export default classes;
+}
