+++ frontend/package.json
@@ -25,14 +25,13 @@
     "react-redux": "^9.2.0",
     "react-router-dom": "^7.6.0",
     "react-scripts": "5.0.1",
-    "styled-components": "^6.1.18",
+    "sass": "^1.88.0",
     "typescript": "^4.4.2",
     "web-vitals": "^2.1.0",
     "yup": "^1.6.1"
   },
   "devDependencies": {
     "@types/react-router-dom": "^5.3.3",
-    "@types/styled-components": "^5.1.34",
     "babel-jest": "^29.7.0",
     "jest": "^29.7.0",
     "ts-jest": "^29.3.2"
