+++ frontend/src/components/login/loginComponent.module.scss
@@ -1,69 +1,67 @@
-import styled from 'styled-components';
-
-export const LoginSection = styled.section`
+.loginSection {
   display: flex;
   align-items: center;
   justify-content: center;
   min-height: 100vh;
-`;
+}
 
-export const LoginContainer = styled.div`
+.loginContainer {
   display: flex;
   flex-direction: row;
 
   @media (max-width: 768px) {
     flex-direction: column;
   }
-`;
+}
 
-export const ImageWrapper = styled.div`
+.imageWrapper {
   flex: 1;
   display: flex;
   align-items: center;
   justify-content: center;
   padding: 2rem;
-`;
+}
 
-export const LoginImage = styled.img`
+.loginImage {
   width: 80%;
   max-width: 400px;
   height: auto;
   object-fit: cover;
-`;
+}
 
-export const FormWrapper = styled.div`
+.formWrapper {
   flex: 1;
   padding: 2rem;
   display: flex;
   flex-direction: column;
   justify-content: center;
-`;
+}
 
-export const FormTitle = styled.h2`
+.formTitle {
   font-size: 1.8rem;
   margin-bottom: 1rem;
   color: #212529;
   text-align: center;
-`;
+}
 
-export const LoginForm = styled.form`
+.loginForm {
   display: flex;
   flex-direction: column;
   gap: 1rem;
-`;
+}
 
-export const FormGroup = styled.div`
+.formGroup {
   display: flex;
   flex-direction: column;
-`;
+}
 
-export const Label = styled.label`
+.label {
   font-size: 1rem;
   color: #212529;
   margin-bottom: 0.25rem;
-`;
+}
 
-export const Input = styled.input`
+.input {
   padding: 0.75rem;
   font-size: 1rem;
   border: 1px solid #ced4da;
@@ -75,9 +73,9 @@ export const Input = styled.input`
     border-color: #6c63ff;
     box-shadow: 0 0 0 2px rgba(108, 99, 255, 0.2);
   }
-`;
+}
 
-export const LoginButton = styled.button`
+.loginButton {
   background: #6c63ff;
   color: #fff;
   font-size: 1rem;
@@ -91,16 +89,16 @@ export const LoginButton = styled.button`
   &:hover {
     background: #5b54d3;
   }
-`;
+}
 
-export const BottomText = styled.p`
+.bottomText {
   font-size: 0.9rem;
   color: #6c757d;
   margin-top: 1rem;
   text-align: center;
-`;
+}
 
-export const RegisterLink = styled.a`
+.registerLink {
   color: #6c63ff;
   text-decoration: none;
   font-weight: 500;
@@ -109,4 +107,4 @@ export const RegisterLink = styled.a`
   &:hover {
     text-decoration: underline;
   }
-`;
+}
