+++ frontend/src/pages/login/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import LoginComponent from '../../components/login/index';
+import LoginComponent from '../../components/login/loginComponent';
 
 const LoginPage: React.FC = () => {
     return <LoginComponent  />
